refactor(location): change location in a `InputsTodo` file and change imports
