+++ src/app/todo/Todo.tsx
@@ -3,7 +3,7 @@
 import { useContext } from "react";
 import { TodoContextType } from "@/types/ITodo";
 import { TodoContext } from "@/app/context/todoContext";
-import InputsTodo from "../components/InputsTodo";
+import InputsTodo from "../components/inputs/InputsTodo";
 import RenderAllTodos from "../components/RenderAllTodos";
 
 export default function Todo() {
